Modified averaging for radiation levels, RDA5807M seek workaround
Changed the averaging for the radiation level from a one second rolling average with 20 buckets to a 10 second rolling average with 50 buckets.  The LED color is much more stable now, but is still fairly responsive.

Updated the CPM to uSv/hr ratio to reflect real-world comparison with a calibrated instrument

The RDA5807M seek function doesn't consistently stop on stations.  Instead, we're just going to set a small frequency range and make the button step through the range in 200kHz increments.